refactor(debug_server): extract /reboot + /sdcard + /widget admin family to debug_server_admin (refs #332)
Wave 23b · 15th per-family extract.  System-control + storage +
widget-injection handlers move from debug_server.c into a new
sibling module debug_server_admin.{c,h}:

  POST /reboot   — esp_restart after a 500 ms grace
  GET  /sdcard   — mount state + total/free MB + root + /sdcard/rec
                   directory listings (uses _list_dir_to_json helper)
  POST /widget   — test-harness hook that injects a widget_live (or
                   dismisses one) into the widget store, bypassing
                   Dragon.  Body matches widget_live schema (TinkerBox
                   docs/protocol.md §17.2).  Used by Widget Platform
                   tests before the Dragon Tab5Surface is wired.

Helpers moved alongside:
  - _list_dir_to_json (renamed list_dir_to_json — not exposed)
  - async_widget_refresh (LVGL-thread re-render after upsert/dismiss)

E2E verification on physical Tab5 192.168.1.90:
  - story_smoke 14/14 in 75s
  - GET /sdcard  → mounted=true, total_mb=121942, free_mb=121730, file listing
  - POST /widget {"card_id":"test_admin",...} → {"ok":true}
  - /reboot deliberately not exercised (would terminate test)

debug_server.c LOC: 1,907 → ~1,640 (-267).
debug_server_admin.c: 0 → 270.
